requirements.txt
alembic==1.4.2
aniso8601==8.0.0
asgiref==3.2.7
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
configparser==5.0.0
Django==3.0.7
django-request==1.5.5
docopt==0.6.2
Flask==1.1.2
Flask-Jsonpify==1.5.0
Flask-JWT==0.3.2
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-pymysql==0.2.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
marshmallow-sqlalchemy-pk==0.13.2
mysqlclient==2.0.1
pipreqs==0.4.10
pycparser==2.20
PyJWT==1.4.2
PyMySQL==0.10.0
PyMysqlDB==0.0.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
urllib3==1.25.10
Werkzeug==1.0.1
yarg==0.1.9
